Add a Console Log dialog box
 ! Current annoyances: the console log does not update while the game is
   running and it doesn't properly display console color codes
 * Changed the initial window size to the display size
 + Added a toggle-able dialog box for console output
    ! The build actions in the menubar and toolbar will show the console log
 * Fixed startup to change to the script directory
 * Moved subprocess calls in the Compiler to the Console
    ! While a command is running in the Console, no other windows will update
 - Removed .beef from the git lfs tracking list
 + Added .wav to the git lfs tracking list
 * Fixed Editor changing and unsaved setting by binding to the proper event
 * Made a resource-specific deletion function
 * Moved Object initial event code generation to its own function
 + Added console log toggling to the menubar and toolbar

+++ resources/base.py
@@ -196,7 +196,7 @@ class BEEFBaseResource:
 		self.inputs[name] = BEEFEditor(self.page)
 		ed = self.inputs[name]
 
-		self.page.Bind(wx.stc.EVT_STC_CHARADDED, self.onEditor, ed)
+		self.page.Bind(wx.stc.EVT_STC_CHANGE, self.onEditor, ed)
 
 		return ed
 	def pageMakeGrid(self, name):
@@ -397,11 +397,18 @@ class BEEFBaseResource:
 
 		self.destroyPage()
 
-		self.resourceList[self.id] = None
-		oldfile = self.top.rootDir+self.path+self.name+".json"
-		if os.path.isfile(oldfile):
-			os.remove(oldfile)
+		del self.resourceList[self.id]
+		self.MenuDeleteSpecific(event)
 
 		self.top.treectrl.Delete(self.treeitem)
 	def MenuDuplicate(self, event):
 		pass # See implementation in specific resource file
+
+	def MenuDeleteSpecific(self, event):
+		oldfile = self.top.rootDir+self.path+self.name+".json"
+		if os.path.isfile(oldfile):
+			os.remove(oldfile)
+		if "path" in self.properties:
+			oldfile = self.top.rootDir+self.properties["path"]
+			if os.path.isfile(oldfile):
+				os.remove(oldfile)
